I fixed some of the issues
the primary problem is that some of the variables were not being pulled properly so I went and changed some. An example of this is beach_learn_about_study used to pull how they heard about us rather than beachphone_hear_about_us. 

there are still some issues in that I can't pull any dates. this is presumably an access problem so we'll have to move forward without that information for now.